Create separate SUR shared folder
